update vendor directory to 1.4 origin
This is an experimental proof of concept, it can use some refinement
before merging. I am proposing this to help with the effort to
understand the issues related to replication controller updates getting
scaled back.

+++ tests/unit/services_test.go
@@ -57,8 +57,8 @@ func (s *OshinkoUnitTestSuite) TestServicePorts(c *check.C) {
 }
 
 func (s *OshinkoUnitTestSuite) TestServicePort(c *check.C) {
-	expectedPort := 12345
-	newServicePort := services.ServicePort(expectedPort)
+	expectedPort := int32(12345)
+	newServicePort := services.ServicePort(int(expectedPort))
 	c.Assert(newServicePort.ServicePort.Port, check.Equals, expectedPort)
 }
